More movement & cleanup
